optimizo winesApiController

+++ backend/src/controllers/winesApiController.js
@@ -1,12 +1,12 @@
 const db = require("../database/models");
+
 const winesApiController = {
     findAll: async (req, res) => {
         const vinos = await db.Vinos.findAll({
             include: [{ all: true }],
             order: [["nombre", "ASC"]],
         });
-        return res.json({
-            status: 200,
+        res.status(200).json({
             total: vinos.length,
             data: vinos,
         });
@@ -19,82 +19,78 @@ const winesApiController = {
         });
 
         if (!vino) {
-            return res.status(400).json({
-                status: 400,
+            res.status(404).json({
                 error: "No contamos con ese vino",
             });
+        } else {
+            res.status(200).json({
+                data: vino,
+            });
         }
-
-        return res.json({
-            status: 200,
-            data: vino,
-        });
     },
 
     create: async (req, res) => {
-        const vino = await db.Vinos.create({
-            nombre: req.body.nombre,
-            //imagen: req.file.path.split("public").pop(),
-            imagen: req.body.imagen,
-            bodega_id: req.body.bodega_id,
-            descripcion: req.body.descripcion,
-            precio: req.body.precio,
-            anio: req.body.anio,
-            uva_id: req.body.uva_id,
-            categoria_id: req.body.categoria_id,
-            stock: req.body.stock,
+        const yaExiste = await db.Vinos.findOne({
+            include: [{ all: true }],
+            where: { nombre: req.body.nombre },
         });
-        return res.status(200).json({
-            status: 200,
-            created: "OK",
-            data: vino,
-        });
+
+        if (yaExiste) {
+            res.status(400).json({
+                error: "Ya contamos con ese vino en nuestra DB",
+                data: yaExiste,
+            });
+        } else {
+            const newWine = await db.Vinos.create({ ...req.body }); //probablemente tenga que modificar la propiedad imagen del body
+            res.status(200).json({
+                created: "OK",
+                data: newWine,
+            });
+        }
     },
 
     modify: async (req, res) => {
+        const vinoModificable = await db.Vinos.findOne({
+            include: [{ all: true }],
+            where: { id: req.params.id },
+        });
+
+        if (!vinoModificable) {
+            res.status(404).json({
+                error: "No contamos con ese vino",
+            });
+        } else {
+            res.status(200).json({
+                data: vinoModificable,
+            });
+        }
+    },
+
+    update: async (req, res) => {
         const vino = await db.Vinos.findOne({
             include: [{ all: true }],
             where: { id: req.params.id },
         });
 
         if (!vino) {
-            return res.status(400).json({
-                status: 400,
-                error: "No contamos con ese vino",
+            res.status(404).json({
+                error: "No se pudo modificar un vino inexistente",
+            });
+        } else {
+            await db.Vinos.update(
+                { ...req.body },
+                {
+                    where: {
+                        id: req.params.id,
+                    },
+                }
+            );
+            res.status(200).json({
+                updated: "OK",
+                updatedData: req.body,
+                oldData: vino,
             });
         }
-
-        return res.json({
-            status: 200,
-            data: vino,
-        });
-    },
-
-    update: async (req, res) => {
-        const vino = await db.Vinos.update(
-            {
-                nombre: req.body.nombre,
-                //imagen: req.file.path.split("public").pop(),
-                imagen: req.body.imagen,
-                bodega_id: req.body.bodega_id,
-                descripcion: req.body.descripcion,
-                precio: req.body.precio,
-                anio: req.body.anio,
-                uva_id: req.body.uva_id,
-                categoria_id: req.body.categoria_id,
-                stock: req.body.stock,
-            },
-            {
-                where: {
-                    id: req.params.id,
-                },
-            }
-        );
-        return res.status(200).json({
-            status: 200,
-            updated: "OK",
-            data: vino,
-        });
     },
 
     delete: async (req, res) => {
@@ -104,29 +100,37 @@ const winesApiController = {
         });
 
         if (!vino) {
-            return res.status(400).json({
-                status: 400,
+            res.status(404).json({
                 error: "No contamos con ese vino",
             });
+        } else {
+            res.status(200).json({
+                data: vino,
+            });
         }
-
-        return res.json({
-            status: 200,
-            data: vino,
-        });
     },
 
     destroy: async (req, res) => {
-        const vino = await db.Vinos.destroy({
-            where: {
-                id: req.params.id,
-            },
+        const vinoAEliminar = await db.Vinos.findOne({
+            include: [{ all: true }],
+            where: { id: req.params.id },
         });
-        return res.status(200).json({
-            status: 200,
-            deleted: "OK",
-            data: vino,
-        });
+
+        if (vinoAEliminar) {
+            await db.Vinos.destroy({
+                where: {
+                    id: req.params.id,
+                },
+            });
+            res.status(200).json({
+                deleted: "OK",
+                deletedWine: vinoAEliminar,
+            });
+        } else {
+            res.status(400).json({
+                error: "No se pudo eliminar un vino inexistente",
+            });
+        }
     },
 };
